+++ objective-c/lessons-from_teach-yourself-objective-c/Simple Mac App/Simple Mac App/AppDelegate.h
@@ -1,0 +1,18 @@
+//
+//  AppDelegate.h
+//  Simple Mac App
+//
+//  Created by Juan Gallardo on 4/9/13.
+//  Copyright (c) 2013 Juan Is Awesome. All rights reserved.
+//  blah blah blah blah
+
+#import <Cocoa/Cocoa.h>
+
+@interface AppDelegate : NSObject <NSApplicationDelegate>
+
+
+//  super important comment goes here
+
+@property (assign) IBOutlet NSWindow *window;
+
+@end