+++ objective-c/lessons-from_teach-yourself-objective-c/Simple Mac App/Simple Mac App/AppDelegate.m
@@ -1,0 +1,18 @@
+//
+//  AppDelegate.m
+//  Simple Mac App
+//
+//  Created by Juan Gallardo on 4/9/13.
+//  Copyright (c) 2013 Juan Is Awesome. All rights reserved.
+//
+
+#import "AppDelegate.h"
+
+@implementation AppDelegate
+
+- (void)applicationDidFinishLaunching:(NSNotification *)aNotification
+{
+	// Insert code here to initialize your application
+}
+
+@end