+++ objective-c/lessons-from_teach-yourself-objective-c/Simple Mac App/Simple Mac App/main.m
@@ -1,0 +1,14 @@
+//
+//  main.m
+//  Simple Mac App
+//
+//  Created by Juan Gallardo on 4/9/13.
+//  Copyright (c) 2013 Juan Is Awesome. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+int main(int argc, char *argv[])
+{
+	return NSApplicationMain(argc, (const char **)argv);
+}