+++ objective-c/lessons-from_teach-yourself-objective-c/first/first/main.m
@@ -1,0 +1,18 @@
+//
+//  main.m
+//  first
+//
+//  Created by Juan Gallardo on 4/9/13.
+//  Copyright (c) 2013 Juan Is Awesome. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+#import "tyoAppDelegate.h"
+
+int main(int argc, char *argv[])
+{
+	@autoreleasepool {
+	    return UIApplicationMain(argc, argv, nil, NSStringFromClass([tyoAppDelegate class]));
+	}
+}