+++ objective-c/lessons-from_teach-yourself-objective-c/first/first/tyoDetailViewController.h
@@ -1,0 +1,16 @@
+//
+//  tyoDetailViewController.h
+//  first
+//
+//  Created by Juan Gallardo on 4/9/13.
+//  Copyright (c) 2013 Juan Is Awesome. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface tyoDetailViewController : UIViewController
+
+@property (strong, nonatomic) id detailItem;
+
+@property (weak, nonatomic) IBOutlet UILabel *detailDescriptionLabel;
+@end